Revert "Move pages and cookie preferences to Publish namespace"
This reverts commit 9e1dfb56aa866e35b4db258e79d428606105288d.
