+++ app/controllers/cookie_preferences_controller.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class CookiePreferencesController < ApplicationController
+  skip_before_action :authenticate
+end
