+++ app/controllers/publish/application_controller.rb
@@ -2,7 +2,6 @@
 
 module Publish
   class ApplicationController < ::ApplicationController
-    include Authentication
     include SuccessMessage
 
     before_action :check_interrupt_redirects
