+++ app/views/cookie_preferences/show.html.erb
