+++ app/views/pages/add_schools_and_study_sites.html.erb
