+++ app/views/pages/course_summary_examples.erb
