+++ app/views/pages/help_writing_course_descriptions.html.erb
