+++ app/views/pages/performance_dashboard.html.erb
