+++ app/views/pages/performance_dashboard/_courses_tab.html.erb
