+++ app/views/pages/performance_dashboard/_initial_performance_indicators.html.erb
