+++ app/views/pages/performance_dashboard/_providers_tab.html.erb
