+++ app/views/pages/performance_dashboard/_rollover_tab.html.erb
