+++ app/views/pages/performance_dashboard/_users_tab.html.erb
