+++ app/views/pages/privacy.html.erb
