+++ app/views/pages/terms.html.erb
